Add match syntax to work with Result of reading file
In order to start accessing the elements of the contents of
/proc/loadavg, it was necessary to employ match syntax so that
it would be a String, that could be coerced to a &str.

Programme now returns the cotents of /proc/loadavg at the
moment of execution. A loop {} was added to main(); however,
it just repeats the value at the time of execution, not the
most recent.  Will need to figure out how to end the file
read and do it again, I think.