Add pattern matching for day part of date
Added pattern a match statement to so as to enable the get_time()
function to print a day of the week (e.g., Monday, Tuesday, etc.).

This may be better as its own function that takes a time object and
returns a &str.

I will follow this with similar functionality for months.